Changes to the save system
Save system was updated. It will work both as the "trainer card" and save menu.

You now need to press A to Save, B returns to the Start Menu. In the future, instead of choosing between Yes and No, I'll make it so pressing A will save with no prompt.

Money was also added to the Save Menu.

Debug settings were slightly altered for testing purposes.

Still need to change the Continue Menu's position.