Talk System
Added the talk system.

This system adds a menu before a trainer battle, allowing the player to talk to the trainers. It displays the Start Battle text.

For now, the system is fully working, with a working menu based on the battle menu and a routine to save the start battle text similar to the one used to save the end battle text.

The Run button is not working properly, it just starts the battle and need to be fixed.

Text will need to be changed, as well as trainer classes. Also, there won't be text in-map anymore, just the exclamation mark. This will be changed in the future.